openhab/scripts/requirements.txt
websockets
requests
numpy
keras-facenet
tensorflow
python-dotenv